AR(1) process to WV

This function computes the Haar WV of an AR(1) process

ar1_to_wv(phi, sigma2, tau)

Arguments

phi

A double that is the phi term of the AR(1) process

sigma2

A double corresponding to variance of AR(1) process

tau

A vec containing the scales e.g. \(2^{\tau}\)

Value

A vec containing the wavelet variance of the AR(1) process.

Details

This function is significantly faster than its generalized counter part arma_to_wv.

Process Haar Wavelet Variance Formula

The Autoregressive Order \(1\) (AR(\(1\))) process has a Haar Wavelet Variance given by: $$\frac{{2{\sigma ^2}\left( {4{\phi ^{\frac{{{\tau _j}}}{2} + 1}} - {\phi ^{{\tau _j} + 1}} - \frac{1}{2}{\phi ^2}{\tau _j} + \frac{{{\tau _j}}}{2} - 3\phi } \right)}}{{{{\left( {1 - \phi } \right)}^2}\left( {1 - {\phi ^2}} \right)\tau _j^2}}$$

See also