ARMA(1,1) to WV

This function computes the WV (haar) of an Autoregressive Order 1 - Moving Average Order 1 (ARMA(1,1)) process.

arma11_to_wv(phi, theta, sigma2, tau)

Arguments

phi
A double corresponding to the autoregressive term.

theta
A double corresponding to the moving average term.

sigma2
A double the variance of the process.

tau
A vec containing the scales e.g. \(2^{\tau}\)

Value

A vec containing the wavelet variance of the ARMA(1,1) process.

Details

This function is significantly faster than its generalized counter part arma_to_wv

Process Haar Wavelet Variance Formula

The Autoregressive Order \(1\) and Moving Average Order \(1\) (ARMA(\(1\),\(1\))) process has a Haar Wavelet Variance given by: $$\nu _j^2\left( {\phi ,\theta ,{\sigma ^2}} \right) = - \frac{{2{\sigma ^2}\left( { - \frac{1}{2}{{(\theta + 1)}^2}\left( {{\phi ^2} - 1} \right){\tau _j} - (\theta + \phi )(\theta \phi + 1)\left( {{\phi ^{{\tau _j}}} - 4{\phi ^{\frac{{{\tau _j}}}{2}}} + 3} \right)} \right)}}{{{{(\phi - 1)}^3}(\phi + 1)\tau _j^2}}$$

See also