Calculate Theoretical Allan Variance for Stationary White Noise Process

This function allows us to calculate the theoretical allan variance for stationary white noise process.

av_wn(sigma2, n)

Arguments

sigma2

A double value for the variance parameter \(\sigma ^2\).

n

An integer value for the size of the cluster.

Value

A double indicating the theoretical allan variance for the white noise process.

Note

This function is based on the calculation of the theoretical allan variance for stationary white noise process raised in "Allan Variance of Time Series Models for Measurement Data" by Nien Fan Zhang. This calculation is fundamental and necessary for the study in "A Study of the Allan Variance for Constant-Mean Non-Stationary Processes" by Xu et al. (IEEE Signal Processing Letters, 2017).

Author

Yuming Zhang

Examples

av1 = av_wn(sigma2 = 1, n = 5)
av2 = av_wn(sigma2 = 2, n = 8)