Computes the MO/DWT wavelet variance for multiple processes

Calculates the MO/DWT wavelet variance

batch_modwt_wvar_cpp(signal, nlevels, robust, eff, alpha, ci_type, strWavelet,
  decomp)

Arguments

signal
A matrix that contains the same number of observations per dataset

robust
A boolean that triggers the use of the robust estimate.

eff
A double that indicates the efficiency as it relates to an MLE.

alpha
A double that indicates the \(\left(1-p\right)\times \alpha\) confidence level

ci_type
A string indicating the confidence interval being calculated. Valid value: "eta3"

strWavelet
A string indicating the type of wave filter to be applied. Must be "haar"

decomp
A string indicating whether to use "modwt" or "dwt" decomp

Value

A field<mat> with the structure:

"variance"Wavelet Variance

"low"Lower CI

"high"Upper CI

Details

This function processes the decomposition of multiple signals quickly

Examples

x = cbind(rnorm(100),rnorm(100))
batch_modwt_wvar_cpp(x, nlevels=4, robust=FALSE, eff=0.6,
                     alpha = 0.05, ci_type="eta3", strWavelet="haar",
                     decomp="modwt")

#>      [,1]      
#> [1,] Numeric,12
#> [2,] Numeric,12