Generate a Confidence interval for a Univariate Time Series

Computes an estimate of the multiscale variance and a chi-squared confidence interval

ci_wave_variance(signal_modwt_bw, wv, type = "eta3", alpha_ov_2 = 0.025,
  robust = FALSE, eff = 0.6)

Arguments

signal_modwt_bw
A field<vec> that contains the modwt or dwt decomposition

wv
A vec that contains the wave variance.

type
A String indicating the confidence interval being calculated.

alpha_ov_2
A double that indicates the \(\left(1-p\right)*\alpha\) confidence level.

robust
A boolean to determine the type of wave estimation.

eff
A double that indicates the efficiency.

Value

A matrix with the structure:

Column 1Wavelet Variance

Column 2Chi-squared Lower Bounds

Column 3Chi-squared Upper Bounds

Details

This function can be expanded to allow for other confidence interval calculations.

Examples

set.seed(1337)
x = rnorm(100)
# Uses the internal MODWT function not associated with an S3 class.
decomp = modwt_cpp(x, filter_name = "haar", nlevels = 4)
y = wave_variance(decomp)
ci_wave_variance(decomp, y, type = "eta3", alpha_ov_2 = 0.025)

#>            [,1]       [,2]      [,3]
#> [1,] 0.64088561 0.44794042 0.9927384
#> [2,] 0.24456060 0.14944086 0.4714211
#> [3,] 0.14316837 0.07297261 0.3980537
#> [4,] 0.06702458 0.02668282 0.3738411