Create a wvar object

Structures elements into a wvar object

create_wvar(
  obj,
  decomp,
  filter,
  robust,
  eff,
  alpha,
  scales,
  unit,
  mean_diff,
  N,
  ranged,
  J
)

Arguments

obj

A matrix with dimensions N x 3 that contains Wavelet Variance, Lower CI, and Upper CI.

decomp

A string that indicates whether to use a "dwt" or "modwt" decomposition.

filter

A string that specifies the type of wavelet filter used in the decomposition.

robust

A boolean that triggers the use of the robust estimate.

eff

A double that indicates the efficiency as it relates to an MLE.

alpha

A double that specifies the significance level which in turn specifies the \(1-\alpha\) confidence level.

scales

A vec that contains the amount of decomposition performed at each level.

unit

A string that indicates the unit expression of the frequency.

mean_diff

A double that specified the empirical mean of the first difference.

N

A integer that specified the empirical length of the time series.

ranged

A double that specified the scaled range of the data, i.e. (max(x) - min(x))/length(x).

J

A integer that specified the number of scales.

Value

A list with the structure:

"variance": Wavelet variance

"ci_low": Lower CI

"ci_high": Upper CI

"robust": Robust active

"eff": Efficiency level for robust calculation

"alpha": p value used for CI

"unit": String representation of the unit

"mean_diff": Empirical mean of the first difference

"N": Length of the time series

"ranged": Scaled range of the data, i.e. (max(x) - min(x))/length(x)

"J": Number of scales