Wavelet variance of IMU Data from a KVH1750 IMU sensor

This data set contains wavelet variance of gyroscope and accelerometer data from an KVH1750 sensor.

kvh1750_wv

Format

A list of the following elements:

"sensor": Name of the sensor.

"freq": The frequency at which the error signal is measured.

"n": Sample size of the data.

"type": The types of sensors considered in the data.

"axis": The axes of sensors considered in the data.

"wvar": A list containing the computed wavelet variance based on the data.

Source

The IMU data comes from Department of Geomatics Engineering, University of Calgary.