Maximum Overlap Discrete Wavelet Transform

Calculation of the coefficients for the discrete wavelet transformation

modwt_cpp(x, filter_name, nlevels)

Arguments

x
A vector with dimensions \(N\times 1\).

filter_name
A string indicating the filter.

nlevels
An integer, \(J\), indicating the level of the decomposition.

Value

y A field<vec> that contains the wavelet coefficients for each decomposition level

Details

Performs a level J decomposition of the time series using the pyramid algorithm. Use this implementation to supply custom parameters instead of modwt(x), which serves as a wrapper function.