Cross Covariance of Matrix

Calculates the Cross-covariance between multiple wavelet transfomations (dwt or modwt)

wccv(x, decomp = "modwt", filter = "haar", nlevels = NULL)

Arguments

x
A vector with dimensions N x M.

decomp
A string that indicates whether to use the "dwt" or "modwt" decomposition.

filter
A string that specifies what wavelet filter to use.

nlevels
An integer that indicates the level of decomposition. It must be less than or equal to floor(log2(length(x))).

Value

Returns a matrix of lists of all the possible pair cross-covariance, variance of each wavelet cross-covariance and its 95

Details

If nlevels is not specified, it is set to \(\left\lfloor {{{\log }_2}\left( {length\left( x \right)} \right)} \right\rfloor\)