Computes the (MODWT) wavelet variance

Calculates the (MODWT) wavelet variance

wvar_cpp(signal_modwt_bw, robust, eff, alpha, ci_type)

Arguments

signal_modwt_bw
A field<vec> that contains the modwt decomposition after it has been brick walled.

robust
A boolean that triggers the use of the robust estimate.

eff
A double that indicates the efficiency as it relates to an MLE.

alpha
A double that indicates the \(\left(1-p\right)*\alpha\) confidence level

ci_type
A String indicating the confidence interval being calculated. Valid value: "eta3"

Value

A mat with the structure:

"variance"Wavelet Variance

"low"Lower CI

"high"Upper CI

Details

This function does the heavy lifting with the signal_modwt_bw

Examples

x = rnorm(100)
decomp = modwt_cpp(x, filter_name = "haar", nlevels = 4)
wvar_cpp(decomp, robust=FALSE, eff=0.6, alpha = 0.05, ci_type="eta3")

#>            [,1]       [,2]      [,3]
#> [1,] 0.40406604 0.28241781 0.6259024
#> [2,] 0.18231825 0.11140713 0.3514412
#> [3,] 0.10522052 0.05363067 0.2925466
#> [4,] 0.03471457 0.01382005 0.1936265