Create a wvar object

Structures elements into a wvar object

create_wvar(obj, decomp, filter, robust, eff, alpha, scales, unit)

Arguments

obj
A matrix with dimensions N x 3, that contains the wavelet variance, low ci, hi ci.

decomp
A string that indicates whether to use the "dwt" or "modwt" decomposition

filter
A string that specifies the type of wavelet filter used in the decomposition

robust
A boolean that triggers the use of the robust estimate.

eff
A double that indicates the efficiency as it relates to an MLE.

alpha
A double that indicates the \(\left(1-p\right)*\alpha\) confidence level

scales
A vec that contains the amount of decomposition done at each level.

unit
A string that contains the unit expression of the frequency.

Value

A list with the structure:

"variance"

Wavelet Variance

"ci_low"

Lower CI

"ci_high"

Upper CI

"robust"

Robust active

"eff"

Efficiency level for Robust

"alpha"

p value used for CI

"unit"

String representation of the unit